+++ eggs/naaya.content.meeting/naaya/content/meeting/eionet_survey/eionet_survey.py
@@ -8,142 +8,296 @@ EIONET_SURVEYS = {
                         'EEA/Eionet cooperation. This evaluation form covers '
                         'all meetings in 2015.'),
         'questions': [
-            {'meta_type': 'Naaya Checkboxes Widget',
+            {'meta_type': 'Naaya Label Widget',
+             'title': 'Introductory Questions',
+             'sortorder': 1},
+
+            {'meta_type': 'Naaya Radio Widget',
              'title': 'I am attending the NFP/Eionet meeting as',
-             'choices': ['NFP (including NFP alternates and assistants)',
+             'choices': ['NFP',
+                         'NFP alternate or assistant',
                          'ETC',
-                         'Other Participant'],
-             'sortorder': 1,
-             'required': False},
-            {'meta_type': 'Naaya Radio Widget',
-             'title': 'I am attending the NFP/Eionet meeting',
-             'choices': ['For the first time',
-                         '2-6 times',
-                         '7-10 times',
-                         'More than 10 times'],
+                         'European Commission'],
              'sortorder': 2,
+             'add_extra_choice': True,
              'required': False},
-            {'meta_type': 'Naaya Radio Matrix Widget',
-             'title': '1. General',
-             'choices': ['Poor', 'Satisfactory', 'Good', 'Excellent',
-                         'I prefer not to answer'],
-             'rows': [('1.1. How do you rate this NFP/Eionet meeting in '
-                       'terms of usefulness?'),
-                      ('1.2. How do you rate this NFP/Eionet meeting in '
-                       'terms of organisation?')],
+
+            {'meta_type': 'Naaya String Widget',
+             'title': 'I work at',
+             'tooltips': '(your host organisation and country)',
              'sortorder': 3,
-             'required': False},
-            {'meta_type': 'Naaya Radio Matrix Widget',
-             'title': '2. Meeting agenda and objectives',
-             'choices': ['Poor', 'Satisfactory', 'Good', 'Excellent',
-                         'I prefer not to answer'],
-             'rows': [('2.1 Are the objectives and agenda of the NFP/Eionet '
-                       'meeting clear?'),
-                      '2.2 Are the background materials relevant?',
-                      '2.3 Facilitation / Chairing',
-                      ('2.4 Have the objectives of each session been achieved?'
-                       ),
-                      '2.5 Is the right mixture of topics covered?'],
+             'required': True},
+
+            {'meta_type': 'Naaya Radio Widget',
+             'title': 'Number of NFP/Eionet meetings attended',
+             'choices': ['This was the first meeting',
+                         '2-5 meetings',
+                         '6-10 meetings',
+                         '10 < meetings'],
              'sortorder': 4,
-             'required': False},
-            {'meta_type': 'Naaya Text Area Widget',
-             'title': ('2.6. Additional comments on the scope and clarity of '
-                       'NFP/Eionet meeting agendas'),
+             'required': True},
+
+            {'meta_type': 'Naaya Radio Widget',
+             'title': 'How much time did you spend preparing for the meeting?',
+             'choices': ['0-2 hours',
+                         '2-5 hours',
+                         '5-10 hours',
+                         '10 < hours',
+                         'I did not prepare'],
              'sortorder': 5,
-             'rows': 5},
-            {'meta_type': 'Naaya Radio Matrix Widget',
-             'title': '3. Organisational aspects',
-             'choices': ['Poor', 'Satisfactory', 'Good', 'Excellent',
-                         'I prefer not to answer'],
-             'rows': ['3.1 Travel arrangements prior to meeting',
-                      '3.2 NFP/Eionet secretariat support',
-                      ('3.3 Working conditions during this meeting (conference'
-                       ' room facilities, equipment etc.)')],
+             'required': True},
+
+            {'meta_type': 'Naaya Checkboxes Widget',
+             'title': 'How did you prepare for the meeting?',
+             'choices': ['Read the meeting documents',
+                         'Prepared a presentation, poster, etc. that was '
+                         'shared during the meeting',
+                         'Discussed the meeting documents with my national '
+                         'network (MB member, NRC, other expert)',
+                         'Discussed the meeting documents with other meeting '
+                         'participants',
+                         'Organised a working group meeting'],
              'sortorder': 6,
-             'required': False},
-            {'meta_type': 'Naaya Text Area Widget',
-             'title': ('3.4 Are there any aspects which need extra attention '
-                       'when the NFP/Eionet meeting is held outside the EEA? '
-                       '(2016 = Turkey)'),
-             'sortorder': 7,
-             'rows': 5},
-            {'meta_type': 'Naaya Text Area Widget',
-             'title': ('3.5 Additional comments on organisational aspects of '
-                       'NFP/Eionet meetings overall'),
-             'sortorder': 8,
-             'rows': 5},
+             'required': True},
+
+            {'meta_type': 'Naaya Label Widget',
+             'title': 'Relevance (agenda and content)',
+             'sortorder': 6},
+
             {'meta_type': 'Naaya Radio Matrix Widget',
-             'title': '4. Time allocation',
-             'choices': ['Too short', 'Fine', 'Too long',
-                         'I prefer not to answer'],
-             'rows': [('4.1 Length of time allocated to discussions during '
-                       'the NFP/Eionet meeting'),
-                      '4.2 Time for networking',
-                      ('4.3 Length of time allocated to presentations and '
-                       'lectures during the NFP/Eionet meeting')],
+             'title': ' Please rank accordingly',
+             'tooltips': '5 is the top score',
+             'choices': ['1', '2', '3', '4', '5', 'N/A'],
+             'rows': [
+                 'How relevant was the meeting to your daily work as NFP?',
+                 'How relevant was the meeting to the overall operation of '
+                 'Eionet in your country?',
+                 'How well did the agenda reflect the strategic priorities of '
+                 'EEA/Eionet?',
+                 'How well did the agenda address EU policy issues and '
+                 'processes relevant to Eionet?'
+             ],
+             'sortorder': 7,
+             'required': True},
+
+            {'meta_type': 'Naaya Checkboxes Widget',
+             'title': 'Please select the sessions that were most relevant to '
+                      'your work',
+             'choices': ['Session 1: Governance issues',
+                         'Session 2: Cooperation and Development',
+                         'Session 3: ED outlook',
+                         'Session 4: Integrated Assessments',
+                         'Session 5: 2019 Key Priorities - Plenary: Part A',
+                         'Session 5: 2019 Key Priorities - '
+                         'Break out discussions: Part B',
+                         'Session 5: 2019 Key Priorities - Plenary: Part C',
+                         ],
+             'sortorder': 8,
+             'required': True},
+
+            {'meta_type': 'Naaya Text Area Widget',
+             'title': 'Please give examples of how you intend to use the '
+                      'outcome of this meeting in your daily work as NFP',
              'sortorder': 9,
-             'required': False},
-            {'meta_type': 'Naaya Radio Widget',
-             'title': ('4.4 Do you find it useful to hold the meeting outside'
-                       ' Copenhagen occasionally?'),
-             'choices': ['Yes', 'No', 'I prefer not to answer'],
+             'rows': 5,
+             'required': True},
+
+            {'meta_type': 'Naaya Text Area Widget',
+             'title': 'Please list key topics/issues that you would find '
+                      'relevant in a future NFP/Eionet meeting',
              'sortorder': 10,
-             'required': False},
-            {'meta_type': 'Naaya Text Area Widget',
-             'title': '4.5 Please explain',
-             'sortorder': 11,
-             'rows': 5},
+             'rows': 5,
+             'required': True},
+
             {'meta_type': 'Naaya Label Widget',
-             'title': '5. Opinions on NFP/Eionet meetings',
-             'sortorder': 12},
-            {'meta_type': 'Naaya Text Area Widget',
-             'title': ('5.1 In your opinion, what were the main benefits of '
-                       'this NFP/Eionet meeting?'),
+             'title': 'Performance (chairing and engagement)',
+             'sortorder': 11},
+
+            {'meta_type': 'Naaya Radio Matrix Widget',
+             'title': 'How would you rate the following elements of the '
+                      'meeting?',
+             'tooltips': '5 is the top score',
+             'choices': ['1', '2', '3', '4', '5', 'N/A'],
+             'rows': ['Pre-meeting',
+                      'Facilitation/interaction during break-out sessions',
+                      'Poster session'],
+             'sortorder': 12,
+             'required': True},
+
+            {'meta_type': 'Naaya Radio Matrix Widget',
+             'title': 'How would you rate the chairing of the sessions?',
+             'tooltips': '5 is the top score',
+             'choices': ['1', '2', '3', '4', '5', 'N/A'],
+             'rows': ['Welcome and opening of the meeting',
+                      'S1: Governance issues 	',
+                      'S2: Cooperation and Development',
+                      'S3: ED outlook',
+                      'S4: Integrated Assessments',
+                      'S5: 2019 Key Priorities - Plenary: Part A',
+                      'S5: 2019 Key Priorities - Plenary: Part C',
+                      ],
              'sortorder': 13,
-             'rows': 5},
-            {'meta_type': 'Naaya Text Area Widget',
-             'title': ('5.2  Please cite one session (including discussions) '
-                       'that worked well at this meeting and explain why.'),
+             'required': True},
+
+            {'meta_type': 'Naaya Radio Matrix Widget',
+             'title': 'How would you rate country engagement during this '
+                      'meeting, in terms of',
+             'tooltips': '5 is the top score',
+             'choices': ['1', '2', '3', '4', '5', 'N/A'],
+             'rows': ['General engagement of countries during plenaries',
+                      'General engagement of countries during break out '
+                      'discussions',
+                      'Your contributions to this meeting',
+                      ],
              'sortorder': 14,
-             'rows': 5},
+             'required': True},
+
             {'meta_type': 'Naaya Text Area Widget',
-             'title': ('5.3 Which sessions were less useful? '
-                       'How would you suggest that EEA improves them?'),
+             'title': 'Please elaborate on your ranking above',
              'sortorder': 15,
-             'rows': 5},
-            {'meta_type': 'Naaya Text Area Widget',
-             'title': '6. Additional comments (for example usefulness of joint'
-                      ' meetings with other networks).',
+             'rows': 5,
+             'required': True},
+
+            {'meta_type': 'Naaya Radio Matrix Widget',
+             'title': 'How would you rate the networking opportunities during '
+                      'the meeting, in relation to',
+             'tooltips': '5 is the top score',
+             'choices': ['1', '2', '3', '4', '5', 'N/A'],
+             'rows': ['The Eionet team',
+                      'The Country Desk Officers',
+                      'Other EEA colleagues',
+                      'Other members of Eionet',
+                      'Members of the European Commission',
+                      ],
              'sortorder': 16,
-             'rows': 5},
-            {'meta_type': 'Naaya Label Widget',
-             'title': '7. Opinions on NFP/Eionet webinars',
-             'sortorder': 17},
+             'required': True},
+
             {'meta_type': 'Naaya Text Area Widget',
-             'title': ('7.1 In your opinion, what were the main benefits of '
-                       'NFP/Eionet webinars in 2016?'),
+             'title': 'How would you like networking opportunities to be '
+                      'strengthened in future NFP meetings?',
+             'sortorder': 17,
+             'rows': 5,
+             'required': True},
+
+            {'meta_type': 'Naaya Text Area Widget',
+             'title': 'How would you like to contribute to the success of '
+                      'future NFP meetings?',
              'sortorder': 18,
-             'rows': 5},
-            {'meta_type': 'Naaya Radio Widget',
-             'title': ('7.2 What do you think of the number of NFP webinars '
-                       'per the year? There are:'),
-             'choices': ['Too many', 'The  number is fine', 'Too few',
-                         'I prefer not to answer'],
-             'sortorder': 19,
-             'required': False},
-            {'meta_type': 'Naaya Radio Widget',
-             'title': ('7.3 Do you find the new tools to operate webinar '
-                       'user friendly?'),
-             'choices': ['Yes', 'No', 'I prefer not to answer'],
-             'sortorder': 20,
-             'required': False},
-            {'meta_type': 'Naaya Text Area Widget',
-             'title': ('7.4 Additional comments related to webinars'),
-             'sortorder': 21,
-             'rows': 5},
+             'rows': 5,
+             'required': True},
+
             {'meta_type': 'Naaya Label Widget',
-             'title': 'Thank you!',
-             'sortorder': 22}
+             'title': 'Preparations and Outcome',
+             'sortorder': 19},
+
+            {'meta_type': 'Naaya Radio Matrix Widget',
+             'title': 'How would you rate the preparation of the meeting, '
+                      'in terms of',
+             'tooltips': '5 is the top score',
+             'choices': ['1', '2', '3', '4', '5', 'N/A'],
+             'rows': ['Relevance of background documents',
+                      'Timeliness of background documents',
+                      'Specification of EEA expectations regarding Eionet '
+                      'involvement',
+                      ],
+             'sortorder': 20,
+             'required': True},
+
+            {'meta_type': 'Naaya Radio Matrix Widget',
+             'title': 'How would you rate the outcome of this meeting, in '
+                      'terms of',
+             'tooltips': '5 is the top score',
+             'choices': ['1', '2', '3', '4', '5', 'N/A'],
+             'rows': ['Usefulness',
+                      'Efficiency',
+                      'Effectiveness',
+                      ],
+             'sortorder': 21,
+             'required': True},
+
+            {'meta_type': 'Naaya Radio Matrix Widget',
+             'title': 'To what extent has EEA/Eionet achieved the objectives '
+                      'of this meeting?',
+             'tooltips': '5 is the top score',
+             'choices': ['1', '2', '3', '4', '5', 'N/A'],
+             'rows': ['S1: Governance issues',
+                      'S2: Cooperation and Development',
+                      'S3: ED outlook',
+                      'S4: Integrated Assessments',
+                      'S5: 2019 Key Priorities - Plenary: Part A',
+                      'S5: 2019 Key Priorities - Break out discussions: '
+                      'Part B',
+                      'S5: 2019 Key Priorities - Plenary: Part C',
+                      ],
+             'sortorder': 22,
+             'required': True},
+
+            {'meta_type': 'Naaya Text Area Widget',
+             'title': 'Please explain what could be improved in future '
+                      'meetings',
+             'sortorder': 23,
+             'rows': 5,
+             'required': True},
+
+            {'meta_type': 'Naaya Label Widget',
+             'title': 'Value added',
+             'sortorder': 24},
+
+            {'meta_type': 'Naaya Text Area Widget',
+             'title': 'How would you like EEA/Eionet to work together to '
+                      'implement the work programmes 2019-2020?',
+             'sortorder': 25,
+             'rows': 5,
+             'required': True},
+
+            {'meta_type': 'Naaya Text Area Widget',
+             'title': 'From a national perspective, what do you see as the '
+                      'future challenges for Eionet?',
+             'sortorder': 26,
+             'rows': 5,
+             'required': True},
+
+            {'meta_type': 'Naaya Text Area Widget',
+             'title': 'How would you like to see the role of the NFP develop '
+                      'in the years to come?',
+             'sortorder': 27,
+             'rows': 5,
+             'required': True},
+
+            {'meta_type': 'Naaya Text Area Widget',
+             'title': 'Which main Eionet assets/features would you like to '
+                      'see in the next strategy (2020-2030)?',
+             'sortorder': 28,
+             'rows': 5,
+             'required': True},
+
+            {'meta_type': 'Naaya Text Area Widget',
+             'title': 'Which Eionet features/activities, if any, would you '
+                      'consider outdated/irrelevant for the future of Eionet?',
+             'sortorder': 29,
+             'rows': 5,
+             'required': True},
+
+            {'meta_type': 'Naaya Label Widget',
+             'title': 'Meeting organisation',
+             'sortorder': 30},
+
+            {'meta_type': 'Naaya Radio Matrix Widget',
+             'title': 'How would you rate the following support?',
+             'tooltips': '5 is the top score',
+             'choices': ['1', '2', '3', '4', '5', 'N/A'],
+             'rows': ['EEA secretarial service',
+                      'Travel arrangements provided by BCD',
+                      'Social arrangements',
+                      ],
+             'sortorder': 31,
+             'required': True},
+
+            {'meta_type': 'Naaya Text Area Widget',
+             'title': 'Please list recommendations for improvements',
+             'sortorder': 32,
+             'rows': 5,
+             'required': True},
         ]
     },
     'NRC meeting': {
@@ -195,7 +349,7 @@ EIONET_SURVEYS = {
                       ('3.2 Balance of content between European and country '
                        'level'),
                       '3.3 Facilitation / Chairing (if applicable)',
-                      '3.4 Use of practical examples/ case studies',
+                      '3.4 Use of practical examples / case studies',
                       '3.5 Quality of presentations',
                       '3.6 Engagement with ETC partners, if applicable',
                       '3.7 Exchange of information and experiences'],